update workflows (note that they are obsolute)
