--- workflow/post_experiment_data_processing/config.yaml
@@ -1,56 +1,0 @@
-eddy_calculation_setting:
-  filament:
-    fraction:
-    - 0.3333333333333333
-    - 0.3333333333333333
-    - 0.3333333333333333
-    r:
-    - 0.35
-    - 0.35
-    - 0.35
-    z:
-    - 0.25
-    - 0
-    - -0.25
-  wall_correction:
-    option: 1
-efit_setting:
-  broken: []
-  efit_table_dir: /home/user1/work/efit-new2/build/16_coils/
-  fl_correct_option: 1
-  gaussian_fit_option: 1
-  nffprime: 2
-  npprime: 2
-  run_dir: /home/user1/work/efit-new2/build/efit/efit
-  timeset: auto
-  tstep: 0.001
-  uncertainty:
-  - 1e-4
-  - 1e-4
-  - 5e-2
-  - 3e-2
-  - 1e-2
-  - 1e-1
-  - 1e-2
-  - 1e-1
-  - 1e-2
-  weighting:
-  - 1
-  - 1
-  - 1
-  - 0.1
-  - 0.1
-  - 0.1
-  - 0.01
-  - 0.01
-shots:
-- 39915
-chease_setting:
-  eqdskpy_dir: /home/user1/GPEC/1.5cp/docs/examples/workflow/eqdsk.py
-dcon_setting:
-  stability_analysis_dir: /home/user1/test/laurent/VEST_database/OMAS
-DB_setting:
-  file_name: /srv/vest.filedb/public/VEST_DB.xlsx
-IMAS_setting:
-  version: 3.41.0-5.1.1
-  old_version: 3.32.0
