+++ workflow/running_linear_stability/config.yaml
@@ -1,0 +1,1 @@
+BASE_DIR: "/srv/vest.filedb/public"
